mix : 어항 리스트 기능 일부구현 & 리팩토링
1. 어항 frame생성 최적화
2. 어항 리스트 실시간 변경 사항 파악가능 기능
3. return 값과 api 일치
4. NowDto -> 자료형 int -> double

+++ src/main/java/com/aqualifeplus/aqualifeplus/fishbowl/dto/firebase/NowDTO.java
@@ -8,16 +8,16 @@ import lombok.Data;
 public class NowDTO {
     private boolean co2State;
     private boolean lightState;
-    private int phState;
-    private int tempState;
+    private double phState;
+    private double tempState;
     private int filterState;
 
     public static NowDTO startNowData() {
         return NowDTO.builder()
                 .co2State(false)
                 .lightState(false)
-                .phState(0)
-                .tempState(0)
+                .phState(0.0)
+                .tempState(0.0)
                 .filterState(0)
                 .build();
     }
